native: make the link refuse undefined symbols, and fix the 26 it was hiding
The Linux link passed both -Wl,--unresolved-symbols=ignore-all and
-Wl,--warn-unresolved-symbols. Between them the link could not report a
missing symbol at all: ignore-all suppresses the warn flag, and the warn
flag independently demotes --no-undefined's errors back to warnings (a
--no-undefined build with the warn flag still left in place links fine and
still produces a binary). dc3-native carried 24 undefined symbols with zero
diagnostics.

They are lazily bound, so the link succeeded, the binary booted, and the
process would die with 'symbol lookup error' / exit 127 on the FIRST call.
Confirmed on the pre-fix binary: LD_BIND_NOW=1 ./dc3-native exits 127 on
BinkRegisterFrameBuffers.

Replaces both flags with -Wl,--no-undefined and closes what it surfaced:

  * XNotifyCreateListener - xdk_shims.cpp defined it taking DWORD where
    xbox.h declares QWORD. In C++ that is a new overload, not a definition,
    so it mangled to _Z21XNotifyCreateListenerj while App.cpp kept calling
    the unmangled extern "C" symbol. KinectGuideThread would have died on
    it. Its neighbour XNotifyGetNext was correct, which is why only one of
    the pair was missing.
  * BinkStartAsyncThread - engine_stubs_generated.cpp appears to cover it,
    but via an asm-label alias for _Z20BinkStartAsyncThreadii: C++ mangling
    and the wrong parameter types. Could never have matched the call. Only
    referenced off the HX_FFMPEG path, so only render-test and the export
    tools saw it.
  * 12 more Bink entry points, 3 Kinect NUI, ReadSingleJoypad,
    requestBreedWrite, DmIsDebuggerPresent, and the two BinkMovie platform
    hooks that native/src/platform/BinkMovie_Stub.cpp claims to 'replace'
    while containing only a three-line comment. New file
    native/src/xbox_link_stubs.cpp, following the HX_STUB_TRACE convention.
  * __vmaddfp / __vspltw are implemented for real, not stubbed:
    gesture/Skeleton.cpp transforms joint positions through them, so a
    zero-returning stub would silently corrupt every skeleton rather than
    fail loudly. web_stubs.cpp had them correct but is guarded by
    '#ifdef __EMSCRIPTEN__' and is in no source list, so it is dead twice
    over and the native build never saw them.
  * std::type_info::operator== - not an Xbox symbol. -fms-compatibility
    makes clang stop defining __GXX_WEAK__, which flips libstdc++'s
    __GXX_TYPEINFO_EQUALITY_INLINE to 0 and turns the comparison into an
    out-of-line call that libstdc++.so.6.0.36 does not export. Fixed with
    the override <typeinfo> documents for exactly this.
  * DebugPanel::Toggle() - the engine deliberately does not compile
    DebugPanel.cpp, but libmilo-engine.a's Rnd_Wgpu.cpp calls it. It was
    listed per-target on dc3-native/milo-viewer/dc3-web and missing from
    milo-tests, render-test and all three export tools. Moved into the
    shared engine-linking list so it cannot fall off a target again.

All seven native executables now link with --no-undefined. Remaining
undefined dynamic symbols in dc3-native: __gmon_start__ and the two
_ITM_*TMCloneTable hooks, all WEAK and standard.

scripts/native_test.sh unchanged before and after: 506 registered,
437 executed, 437 passed, 0 failed, 69 skipped (budget 69), exit 0.
Runtime: 900 frames with DC3_STUB_TRACE=1, clean exit, no symbol lookup
errors, and zero hits on any new stub - these were latent on the boot path,
not live.

No src/ file touched, so PPC codegen is unaffected by construction.

Also corrects a stale comment: the export tools are no longer 'unchanged',
and src/platform/Tex_Wgpu.cpp compiles to no object in any native target
(it is byte-identical to the engine's copy, which is what actually links)
while remaining live for dc3-web.

+++ native/src/xdk_shims.cpp
@@ -799,7 +799,17 @@ DWORD XShowNuiDirtyDiscErrorUI(DWORD, DWORD) { exit(1); }
 DWORD XShowDirtyDiscErrorUI(DWORD) { exit(1); }
 DWORD XMarketplaceCreateOfferEnumerator(DWORD, DWORD, ULONGLONG, DWORD, DWORD *, HANDLE *) { return 1; }
 DWORD XMarketplaceCreateOfferEnumeratorByOffering(DWORD, DWORD, ULONGLONG *, WORD, DWORD *, HANDLE *) { return 1; }
-HANDLE XNotifyCreateListener(DWORD) { return nullptr; }
+// NOTE: the parameter is QWORD, not DWORD. xbox.h declares
+//   HANDLE XNotifyCreateListener(QWORD qwAreas);
+// inside an extern "C" block. A definition here with a *different* parameter
+// type is not a redefinition error in C++ -- it is a new overload, and since
+// only one overload can carry C linkage, the DWORD version silently mangled to
+// _Z21XNotifyCreateListenerj while App.cpp's call kept referencing the
+// unmangled extern "C" symbol. The result was an undefined symbol that
+// -Wl,--unresolved-symbols=ignore-all swallowed, leaving KinectGuideThread to
+// die with "symbol lookup error" on first call. Keep this signature in sync
+// with xbox.h.
+HANDLE XNotifyCreateListener(QWORD) { return nullptr; }
 BOOL XNotifyGetNext(HANDLE, DWORD, DWORD *, ULONG_PTR *) { return 0; }
 
 // ============================================================================
